integracion inventario

+++ src/main/java/com/project/sales/entity/SaleDetail.java
@@ -26,7 +26,7 @@ public class SaleDetail {
     private Sale sale;
 
     private BigDecimal price;
-    private Integer cant;
+    private Double cant;
 
 
 }
